Give narrowing one home and mismatches their own sum
Slap.Binary grows the measure-typed vocabulary — takeLength, byteAtOffset, splitSuffixOfLength, fillRegion, seedBufferFromSource, and kin — and offsetToInt, lengthToInt, and fileSizeToInt leave Slap.Measure's exports, so the narrowing into ByteString's Int world happens at one seam instead of 250 call sites across 52 modules. The nine verification mismatches move out of SlapAdvisory into their own VerificationMismatch sum: mismatchClass reads each mismatch's enforcement lane off its constructor, flipMismatchSide is total so a new sided check cannot slip past undo unflipped, and VerificationFatal carries the mismatch itself. Twelve fromIntegral applications that had become the identity are gone, and showSigned takes the Int64 every caller hands it.

+++ src/Slap/BSDiff/Create.hs
@@ -56,9 +56,9 @@ encodeControlStream instructions =
   LazyByteString.toStrict (toLazyByteString (foldMap encodeInstruction instructions))
   where
     encodeInstruction instruction =
-      putSignMagnitude64 (fromIntegral (unLength (controlAdd  instruction)))
-      <> putSignMagnitude64 (fromIntegral (unLength (controlCopy instruction)))
-      <> putSignMagnitude64 (fromIntegral (unDelta  (controlSeek instruction)))
+      putSignMagnitude64 (unLength (controlAdd  instruction))
+      <> putSignMagnitude64 (unLength (controlCopy instruction))
+      <> putSignMagnitude64 (unDelta  (controlSeek instruction))
 
 -- | Write a signed 64-bit LE value in bsdiff sign-magnitude form:
 -- magnitude in the low 63 bits, the sign in bit 63. Round-trip
